added files to modularize the code and updated the output pulling from trails.js and rivers.js with new functions and imports added
