+++ readme.md
@@ -1,3 +1,4 @@
 Debugging exercise for C#
 - added river and trails info.
-- added further details for rivers and trails including metrics for length, least and most expensive tours...
+- added further details for rivers and trails including metrics for length, least and most expensive tours.
+- modularized code to test.
